using sasURL to connect to the azure table insteasd of account key

+++ plugins/outputs/mqtt/mqtt.go
@@ -162,7 +162,7 @@ func (m *MQTT) publish(topic string, body []byte) error {
 
 func (m *MQTT) createOpts() (*paho.ClientOptions, error) {
 	opts := paho.NewClientOptions()
-	opts.KeepAlive = 0 * time.Second
+	opts.KeepAlive = int64(0 * time.Second)
 
 	if m.Timeout.Duration < time.Second {
 		m.Timeout.Duration = 5 * time.Second
